// src/api/init/interfaces/rawdata-ingredient.interface.ts
export interface IRaw_Data_Ingredients {
  ingredient_id: string;
  ingredient: string;
  points: string;
  nb: string;
  note: string;
  avis: string;
  temps: string;
  diff: string;
  cout: string;
  person: string;
  dans_titre: string;
  est_sous_cat: string;
  id_copie: string;
  categorie: string;
  categorie_code: string;
  rank: string;
  total_cat: string;
  ingredient_url: string;
  assimiles: string;
  assimiles_txt: string;
  assimiles_url: string;
  ingredient_code: string;
}